feat(cli): add new update command
gh-528
